Compatibility updates for perccli 007.0529.0000.0000 and PERC H330
Added some fixes for compatibility with PERC H330 and new perccli.
This version of perccli depends on a library called libstorelib.so that is under /opt/lsi/perccli so changed it to cd to that directory or it cannot find the library.
If you connect a disk directly to PERC card then status shows as JBOD so added that to 'good' status list.
The H330 does not have a battery so made that part of the check optional.

Tested with ESXi 6.7u2 with a PERC H330 with perccli 007.0529.0000.0000